UserProfile component setup, moved RegisterForm
