+++ next.config.js
@@ -1,4 +1,7 @@
 module.exports = {
+  images: {
+    domains: ['lh3.googleusercontent.com'],
+  },
   reactStrictMode: true,
   // I don't want it to run when compiling as I trust the CI stage of the pipeline and Husky.
   ignoreDuringBuilds: true,
