--- pages/dashboard.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-function Dashboard() {
-  return (
-    <>
-      <div className="flex flex-col items-center mt-28 text-4xl font-semibold">
-        <h1>Future home of the user dashboard!</h1>
-        <h3>Check back soon!</h3>
-      </div>
-    </>
-  );
-}
-
-export default Dashboard;
