+++ pages/user/[id].js
@@ -1,0 +1,9 @@
+import UserProfile from '../../components/UserProfile';
+
+export default function ViewUserProfile() {
+  // TODO: pass the user ID to the UserProfile component
+
+  return (
+    <UserProfile />
+  );
+}
