+++ utils/auth.js
@@ -2,39 +2,37 @@ import firebase from 'firebase/app';
 import 'firebase/auth';
 import { clientCredentials } from './client';
 
-const checkUser = (uid) =>
-  new Promise((resolve, reject) => {
-    fetch(`${clientCredentials.databaseURL}/api/checkuser`, {
-      method: 'POST',
-      body: JSON.stringify({ uid }),
-      headers: {
-        'Content-Type': 'application/json',
-        Accept: 'application/json',
-      },
+const checkUser = (uid) => new Promise((resolve, reject) => {
+  fetch(`${clientCredentials.databaseURL}/api/checkuser`, {
+    method: 'POST',
+    body: JSON.stringify({ uid }),
+    headers: {
+      'Content-Type': 'application/json',
+      Accept: 'application/json',
+    },
+  })
+    .then((resp) => {
+      if (resp.ok) {
+        resolve(resp.json());
+      } else {
+        resolve({});
+      }
     })
-      .then((resp) => {
-        if (resp.ok) {
-          resolve(resp.json());
-        } else {
-          resolve({});
-        }
-      })
-      .catch(reject);
-  });
+    .catch(reject);
+});
 
-const registerUser = (userInfo) =>
-  new Promise((resolve, reject) => {
-    fetch(`${clientCredentials.databaseURL}/api/register`, {
-      method: 'POST',
-      body: JSON.stringify(userInfo),
-      headers: {
-        'Content-Type': 'application/json',
-        Accept: 'application/json',
-      },
-    })
-      .then((resp) => resolve(resp.json()))
-      .catch(reject);
-  });
+const registerUser = (userInfo) => new Promise((resolve, reject) => {
+  fetch(`${clientCredentials.databaseURL}/api/register`, {
+    method: 'POST',
+    body: JSON.stringify(userInfo),
+    headers: {
+      'Content-Type': 'application/json',
+      Accept: 'application/json',
+    },
+  })
+    .then((resp) => resolve(resp.json()))
+    .catch(reject);
+});
 
 const signIn = () => {
   const provider = new firebase.auth.GoogleAuthProvider();
